debt - move api command things into extHostCommand so that they can be contributed from anywhere
